Remove unnecessary abstract classes
